+++ CHANGELOG.md
@@ -5,3 +5,7 @@
 
 ## [1.0.2] - 2025-01-10
 - Added icons and indents
+
+## [1.0.3] - 2025-01-17
+- arm-none-eabi-objdump and arm-none-eabi-nm are used to get more accurate results
+- symbols now have links to source code